@@ -6,18 +6,22 @@ plugins {
 
 group 'net.sargue'
 archivesBaseName = 'mailgun'
-version '1.10.0'
+version '2.0.0-SNAPSHOT'
 
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(8)
+    }
+}
 
 repositories {
     mavenCentral()
 }
 
 dependencies {
-    implementation 'org.glassfish.jersey.core:jersey-client:2.25'
-    implementation 'org.glassfish.jersey.media:jersey-media-multipart:2.25'
+    implementation 'org.glassfish.jersey.core:jersey-client:3.0.9'
+    implementation 'org.glassfish.jersey.inject:jersey-hk2:3.0.9'
+    implementation 'org.glassfish.jersey.media:jersey-media-multipart:3.0.9'
     testImplementation 'com.github.tomakehurst:wiremock:1.57'
     testImplementation 'org.awaitility:awaitility:2.0.0'
     testImplementation 'org.slf4j:slf4j-simple:1.7.21'
